Add files via upload
This is a beta version of Unfolding Map. 

Unzip the archive and put the extracted Unfolding folder into the libraries folder of your Processing sketches.

To find the Processing sketches location on your computer, open the Preferences window of the Processing application and look for the “Sketchbook location” item at the top. You will need to create the “libraries” folder if this is your first contributed library.

Restart Processing, start a new sketch and create your first simple map.